% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_lasso.R
\name{cv.twin}
\alias{cv.twin}
\title{CV Fitting for A Lasso Model Using the Coordinate Descent Algorithm}
\usage{
cv.twin(x, y, lambda = numeric(0), gamma = 0.25, type.measure = c("mse",
  "deviance", "class", "auc", "mae"), nfolds = 10, foldid = NULL,
  grouped = TRUE, keep = FALSE, parallel = FALSE, ...)
}
\arguments{
\item{x}{The design matrix}

\item{y}{The response vector}

\item{lambda}{A user provided sequence of \eqn{\lambda}. If set to
\code{NULL}, the program will calculate its own sequence
according to \code{nlambda} and \code{lambda_min_ratio},
which starts from \eqn{\lambda_0} (with this
\eqn{\lambda} all coefficients will be zero) and ends at
\code{lambda0 * lambda_min_ratio}, containing
\code{nlambda} values equally spaced in the log scale.
It is recommended to set this parameter to be \code{NULL}
(the default).}

\item{gamma}{bandwidth for two mountains penalty}

\item{type.measure}{measure to evaluate for cross-validation. The default is \code{type.measure = "deviance"},
which uses squared-error for gaussian models (a.k.a \code{type.measure = "mse"} there), deviance for logistic
regression. \code{type.measure = "class"} applies to binomial only. \code{type.measure = "auc"} is for two-class logistic
regression only. \code{type.measure = "mse"} or \code{type.measure = "mae"} (mean absolute error) can be used by all models;
they measure the deviation from the fitted mean to the response.}

\item{nfolds}{number of folds for cross-validation. default is 10. 3 is smallest value allowed.}

\item{foldid}{an optional vector of values between 1 and nfold specifying which fold each observation belongs to.}

\item{grouped}{Like in \pkg{glmnet}, this is an experimental argument, with default \code{TRUE}, and can be ignored by most users.
For all models, this refers to computing nfolds separate statistics, and then using their mean and estimated standard
error to describe the CV curve. If \code{grouped = FALSE}, an error matrix is built up at the observation level from the
predictions from the \code{nfold} fits, and then summarized (does not apply to \code{type.measure = "auc"}).}

\item{keep}{If \code{keep = TRUE}, a prevalidated list of arrasy is returned containing fitted values for each observation
and each value of lambda for each model. This means these fits are computed with this observation and the rest of its
fold omitted. The folid vector is also returned. Default is \code{keep = FALSE}}

\item{parallel}{If TRUE, use parallel foreach to fit each fold. Must register parallel before hand, such as \pkg{doMC}.}

\item{...}{other parameters to be passed to \code{"twin"} function}
}
\description{
Cross validation for linear models with the TWIN penalty

where \eqn{n} is the sample size and \eqn{\lambda} is a tuning
parameter that controls the sparsity of \eqn{\beta}.
}
\examples{
set.seed(123)
n = 100
p = 1000
b = c(runif(10, min = 0.2, max = 1), rep(0, p - 10))
x = matrix(rnorm(n * p, sd = 3), n, p)
y = drop(x \%*\% b) + rnorm(n)

## fit lasso model with 100 tuning parameter values
res <- cv.twin(x, y)


}
