% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinis_methods.R
\name{plot.ordinis}
\alias{plot.ordinis}
\alias{plot.cv.ordinis}
\title{Plot method for ordinis fitted objects}
\usage{
\method{plot}{ordinis}(x, xvar = c("loglambda", "lambda", "penalty"),
  labsize = 0.6, xlab = iname, ylab = NULL, main = x$penalty,
  xlim = NULL, n.print = 10L, ...)

\method{plot}{cv.ordinis}(x, sign.lambda = 1, ...)
}
\arguments{
\item{x}{fitted "ordinis" model object or fitted "cv.ordinis" model object}

\item{xvar}{What is on the X-axis. \code{"penalty"} plots against the penalty value applied to the coefficients, \code{"lambda"} against the log-lambda sequence}

\item{labsize}{size of labels for variable names. If labsize = 0, then no variable names will be plotted}

\item{xlab}{label for x-axis}

\item{ylab}{label for y-axis}

\item{main}{main title for plot}

\item{xlim}{numeric vectors of length 2, giving the \code{x} and \code{y} coordinates ranges.}

\item{n.print}{scalar integer for the number of times along the regularization path to print the number
of nonzero coefficients. If set to a negative value, the number of nonzero coefficients will not be printed.}

\item{...}{other graphical parameters for the plot}

\item{sign.lambda}{Either plot against log(lambda) (default) or its negative if \code{sign.lambda = -1}.}
}
\description{
Plot method for ordinis fitted objects

Plot method for fitted two mountains cv objects
}
\examples{
set.seed(123)
n.obs <- 100
n.vars <- 1000

true.beta <- c(runif(5, 0.1, 1) * (2 * rbinom(5, 1, 0.5) - 1), rep(0, n.vars - 5))

x <- matrix(rnorm(n.obs * n.vars), n.obs, n.vars)
y <- rnorm(n.obs, sd = 2) + x \%*\% true.beta

fit <- ordinis(x = x, y = y, penalty = c("double.quadratic"))

plot(fit)

set.seed(123)
n.obs <- 100
n.vars <- 200

true.beta <- c(runif(15, -0.5, 0.5), rep(0, n.vars - 15))

x <- matrix(rnorm(n.obs * n.vars), n.obs, n.vars)
y <- rnorm(n.obs, sd = 3) + x \%*\% true.beta

fit <- cv.ordinis(x = x, y = y, gamma = 0.1)

plot(fit)

}
