% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinis_methods.R
\name{predict.ordinis}
\alias{predict.ordinis}
\title{Prediction method for coord lasso fitted objects}
\usage{
\method{predict}{ordinis}(object, newx, s = NULL, type = c("link",
  "response", "coefficients", "nonzero", "class"), ...)
}
\arguments{
\item{object}{fitted "ordinis" model object}

\item{newx}{Matrix of new values for \code{x} at which predictions are to be made. Must be a matrix; can be sparse as in the
\code{CsparseMatrix} objects of the \pkg{Matrix} package.
This argument is not used for \code{type=c("coefficients","nonzero")}}

\item{s}{Value(s) of the penalty parameter lambda at which predictions are required. Default is the entire sequence used to create
the model.}

\item{type}{Type of prediction required. \code{type = "link"} gives the linear predictors for the \code{"binomial"} model; for \code{"gaussian"} models it gives the fitted values.
\code{type = "response"} gives the fitted probabilities for \code{"binomial"}. \code{type = "coefficients"} computes the coefficients at the requested values for \code{s}.
\code{type = "class"} applies only to \code{"binomial"} and produces the class label corresponding to the maximum probability.}

\item{...}{not used}
}
\value{
An object depending on the type argument
}
\description{
Prediction method for coord lasso fitted objects
}
\examples{
set.seed(123)
n.obs <- 1e4
n.vars <- 100
n.obs.test <- 1e3

true.beta <- c(runif(15, -0.5, 0.5), rep(0, n.vars - 15))

x <- matrix(rnorm(n.obs * n.vars), n.obs, n.vars)
y <- rnorm(n.obs, sd = 3) + x \%*\% true.beta
x.test <- matrix(rnorm(n.obs.test * n.vars), n.obs.test, n.vars)
y.test <- rnorm(n.obs.test, sd = 3) + x.test \%*\% true.beta

fit <- ordinis(x = x, y = y, nlambda = 10)

preds.lasso <- predict(fit, newx = x.test, type = "response")

apply(preds.lasso, 2, function(x) mean((y.test - x) ^ 2))

}
