% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_lasso_methods.R
\name{logLik.lasso}
\alias{logLik.lasso}
\title{log likelihood function for fitted lasso objects}
\usage{
\method{logLik}{lasso}(object, REML = FALSE, ...)
}
\arguments{
\item{object}{fitted "lasso" model object.}

\item{...}{not used}
}
\description{
log likelihood function for fitted lasso objects
}
\examples{
set.seed(123)
n.obs <- 200
n.vars <- 500

true.beta <- c(runif(15, -0.25, 0.25), rep(0, n.vars - 15))
x <- matrix(rnorm(n.obs * n.vars), n.obs, n.vars)
y <- rnorm(n.obs, sd = 3) + x \%*\% true.beta

fit <- lasso(x = x, y = y)

logLik(fit)


}
